# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Candy Joy Chocolate API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  category: string;
  description: string;
  price: number;
  imageUrl: string;
  featured: boolean;
  createdAt: string;
}

export interface CreateProductBody {
  name: string;
  category: string;
  description: string;
  price: number;
  imageUrl: string;
  featured?: boolean;
}

export interface Review {
  id: number;
  authorName: string;
  rating: number;
  comment: string;
  approved: boolean;
  createdAt: string;
}

export interface CreateReviewBody {
  authorName: string;
  rating: number;
  comment: string;
}

export interface UpdateReviewVisibilityBody {
  approved: boolean;
}

export interface Message {
  id: number;
  name: string;
  email: string;
  message: string;
  read: boolean;
  createdAt: string;
}

export interface CreateMessageBody {
  name: string;
  email: string;
  message: string;
}

export interface MarkMessageReadBody {
  read: boolean;
}

export interface LoginBody {
  username: string;
  password: string;
}

export interface AdminUser {
  username: string;
}

export interface LoginResponse {
  token: string;
  user: AdminUser;
}

export interface StoreSettings {
  id: number;
  phone: string;
  whatsapp: string;
  address: string;
  instagramUrl: string;
  tiktokUrl: string;
  updatedAt: string;
}

export interface UpdateSettingsBody {
  phone?: string;
  whatsapp?: string;
  address?: string;
  instagramUrl?: string;
  tiktokUrl?: string;
}

export interface AdminStats {
  totalProducts: number;
  totalReviews: number;
  totalMessages: number;
  unreadMessages: number;
  averageRating: number;
  recentMessages: Message[];
}

export type ListProductsParams = {
  category?: string;
  search?: string;
  featured?: boolean;
};