# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Candy Joy Chocolate API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product operations
  - name: reviews
    description: Review operations
  - name: messages
    description: Contact message operations
  - name: auth
    description: Authentication operations
  - name: settings
    description: Store settings operations
  - name: admin
    description: Admin stats operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: Get all products
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: featured
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "401":
          description: Unauthorized

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get single product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "401":
          description: Unauthorized
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized

  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: Get all reviews
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "201":
          description: Created review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  /reviews/{id}:
    delete:
      operationId: deleteReview
      tags: [reviews]
      summary: Delete a review (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
    patch:
      operationId: updateReviewVisibility
      tags: [reviews]
      summary: Toggle review visibility (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReviewVisibilityBody"
      responses:
        "200":
          description: Updated review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  /messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: Get all contact messages (admin only)
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: createMessage
      tags: [messages]
      summary: Submit contact form (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMessageBody"
      responses:
        "201":
          description: Created message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  /messages/{id}:
    delete:
      operationId: deleteMessage
      tags: [messages]
      summary: Delete a message (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
    patch:
      operationId: markMessageRead
      tags: [messages]
      summary: Mark message as read (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkMessageReadBody"
      responses:
        "200":
          description: Updated message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials

  /auth/logout:
    post:
      operationId: adminLogout
      tags: [auth]
      summary: Admin logout
      responses:
        "200":
          description: Logged out

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current admin user
      responses:
        "200":
          description: Admin info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Unauthorized

  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get store settings
      responses:
        "200":
          description: Store settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreSettings"
    put:
      operationId: updateSettings
      tags: [settings]
      summary: Update store settings (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSettingsBody"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreSettings"
        "401":
          description: Unauthorized

  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get dashboard stats (admin only)
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
        "401":
          description: Unauthorized

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        description:
          type: string
        price:
          type: number
        imageUrl:
          type: string
        featured:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - category
        - description
        - price
        - imageUrl
        - featured
        - createdAt

    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        description:
          type: string
        price:
          type: number
        imageUrl:
          type: string
        featured:
          type: boolean
      required:
        - name
        - category
        - description
        - price
        - imageUrl

    Review:
      type: object
      properties:
        id:
          type: integer
        authorName:
          type: string
        rating:
          type: integer
        comment:
          type: string
        approved:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - authorName
        - rating
        - comment
        - approved
        - createdAt

    CreateReviewBody:
      type: object
      properties:
        authorName:
          type: string
        rating:
          type: integer
        comment:
          type: string
      required:
        - authorName
        - rating
        - comment

    UpdateReviewVisibilityBody:
      type: object
      properties:
        approved:
          type: boolean
      required:
        - approved

    Message:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        message:
          type: string
        read:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - message
        - read
        - createdAt

    CreateMessageBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        message:
          type: string
      required:
        - name
        - email
        - message

    MarkMessageReadBody:
      type: object
      properties:
        read:
          type: boolean
      required:
        - read

    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password

    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/AdminUser"
      required:
        - token
        - user

    AdminUser:
      type: object
      properties:
        username:
          type: string
      required:
        - username

    StoreSettings:
      type: object
      properties:
        id:
          type: integer
        phone:
          type: string
        whatsapp:
          type: string
        address:
          type: string
        instagramUrl:
          type: string
        tiktokUrl:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - phone
        - whatsapp
        - address
        - instagramUrl
        - tiktokUrl
        - updatedAt

    UpdateSettingsBody:
      type: object
      properties:
        phone:
          type: string
        whatsapp:
          type: string
        address:
          type: string
        instagramUrl:
          type: string
        tiktokUrl:
          type: string

    AdminStats:
      type: object
      properties:
        totalProducts:
          type: integer
        totalReviews:
          type: integer
        totalMessages:
          type: integer
        unreadMessages:
          type: integer
        averageRating:
          type: number
        recentMessages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
      required:
        - totalProducts
        - totalReviews
        - totalMessages
        - unreadMessages
        - averageRating
        - recentMessages