# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Candy Joy Chocolate API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get all products
 */
export const ListProductsQueryParams = zod.object({
  category: zod.coerce.string().optional(),
  search: zod.coerce.string().optional(),
  featured: zod.coerce.boolean().optional(),
});

export const ListProductsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  category: zod.string(),
  description: zod.string(),
  price: zod.number(),
  imageUrl: zod.string(),
  featured: zod.boolean(),
  createdAt: zod.string(),
});
export const ListProductsResponse = zod.array(ListProductsResponseItem);

/**
 * @summary Create a product (admin only)
 */
export const CreateProductBody = zod.object({
  name: zod.string(),
  category: zod.string(),
  description: zod.string(),
  price: zod.number(),
  imageUrl: zod.string(),
  featured: zod.boolean().optional(),
});

/**
 * @summary Get single product
 */
export const GetProductParams = zod.object({
  id: zod.coerce.number(),
});

export const GetProductResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  category: zod.string(),
  description: zod.string(),
  price: zod.number(),
  imageUrl: zod.string(),
  featured: zod.boolean(),
  createdAt: zod.string(),
});

/**
 * @summary Update a product (admin only)
 */
export const UpdateProductParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateProductBody = zod.object({
  name: zod.string(),
  category: zod.string(),
  description: zod.string(),
  price: zod.number(),
  imageUrl: zod.string(),
  featured: zod.boolean().optional(),
});

export const UpdateProductResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  category: zod.string(),
  description: zod.string(),
  price: zod.number(),
  imageUrl: zod.string(),
  featured: zod.boolean(),
  createdAt: zod.string(),
});

/**
 * @summary Delete a product (admin only)
 */
export const DeleteProductParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get all reviews
 */
export const ListReviewsResponseItem = zod.object({
  id: zod.number(),
  authorName: zod.string(),
  rating: zod.number(),
  comment: zod.string(),
  approved: zod.boolean(),
  createdAt: zod.string(),
});
export const ListReviewsResponse = zod.array(ListReviewsResponseItem);

/**
 * @summary Submit a review (public)
 */
export const CreateReviewBody = zod.object({
  authorName: zod.string(),
  rating: zod.number(),
  comment: zod.string(),
});

/**
 * @summary Delete a review (admin only)
 */
export const DeleteReviewParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Toggle review visibility (admin only)
 */
export const UpdateReviewVisibilityParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateReviewVisibilityBody = zod.object({
  approved: zod.boolean(),
});

export const UpdateReviewVisibilityResponse = zod.object({
  id: zod.number(),
  authorName: zod.string(),
  rating: zod.number(),
  comment: zod.string(),
  approved: zod.boolean(),
  createdAt: zod.string(),
});

/**
 * @summary Get all contact messages (admin only)
 */
export const ListMessagesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  message: zod.string(),
  read: zod.boolean(),
  createdAt: zod.string(),
});
export const ListMessagesResponse = zod.array(ListMessagesResponseItem);

/**
 * @summary Submit contact form (public)
 */
export const CreateMessageBody = zod.object({
  name: zod.string(),
  email: zod.string(),
  message: zod.string(),
});

/**
 * @summary Delete a message (admin only)
 */
export const DeleteMessageParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Mark message as read (admin only)
 */
export const MarkMessageReadParams = zod.object({
  id: zod.coerce.number(),
});

export const MarkMessageReadBody = zod.object({
  read: zod.boolean(),
});

export const MarkMessageReadResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  message: zod.string(),
  read: zod.boolean(),
  createdAt: zod.string(),
});

/**
 * @summary Admin login
 */
export const AdminLoginBody = zod.object({
  username: zod.string(),
  password: zod.string(),
});

export const AdminLoginResponse = zod.object({
  token: zod.string(),
  user: zod.object({
    username: zod.string(),
  }),
});

/**
 * @summary Get current admin user
 */
export const GetMeResponse = zod.object({
  username: zod.string(),
});

/**
 * @summary Get store settings
 */
export const GetSettingsResponse = zod.object({
  id: zod.number(),
  phone: zod.string(),
  whatsapp: zod.string(),
  address: zod.string(),
  instagramUrl: zod.string(),
  tiktokUrl: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Update store settings (admin only)
 */
export const UpdateSettingsBody = zod.object({
  phone: zod.string().optional(),
  whatsapp: zod.string().optional(),
  address: zod.string().optional(),
  instagramUrl: zod.string().optional(),
  tiktokUrl: zod.string().optional(),
});

export const UpdateSettingsResponse = zod.object({
  id: zod.number(),
  phone: zod.string(),
  whatsapp: zod.string(),
  address: zod.string(),
  instagramUrl: zod.string(),
  tiktokUrl: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Get dashboard stats (admin only)
 */
export const GetAdminStatsResponse = zod.object({
  totalProducts: zod.number(),
  totalReviews: zod.number(),
  totalMessages: zod.number(),
  unreadMessages: zod.number(),
  averageRating: zod.number(),
  recentMessages: zod.array(
    zod.object({
      id: zod.number(),
      name: zod.string(),
      email: zod.string(),
      message: zod.string(),
      read: zod.boolean(),
      createdAt: zod.string(),
    }),
  ),
});